Update unique_all (tensorflow)

+++ ivy/functional/backends/tensorflow/set.py
@@ -10,14 +10,14 @@ def unique_all(x: Tensor) \
     
     UniqueAll = namedtuple(typename = 'unique_all', field_names = ['values', 'indices', 'inverse_indices', 'counts'])
     
-    flat_tensor = tf.reshape(x, [-1])
+    flat_tensor = tf.cast(tf.reshape(x, [-1]), 'float64')
     
     values, inverse_indices, counts = tf.unique_with_counts(flat_tensor)
     
-    unique_nan = tf.math.is_nan(values).numpy()
+    unique_nan = tf.math.is_nan(tf.cast(values, 'float64')).numpy()
     idx_dtype = inverse_indices.dtype
 
-    if tf.math.reduce_sum(tf.cast(tf.math.is_nan(values), 'int32')).numpy():
+    if tf.math.reduce_sum(tf.math.is_nan(tf.cast(values, 'float64'))).numpy():
         nan_index = tf.where(tf.math.is_nan(flat_tensor)).numpy().reshape([-1])
         non_nan_index = [flat_tensor.numpy().tolist().index(val) for val in values if not tf.math.is_nan(val)]
